Array, Dynamic ADT with Amazing Feature of Resize
a powerful and flexible array implementation that combines the familiar functionalities of a built-in array with enhanced capabilities to meet modern programming needs.
Having all behaviours of built in array of integers.
Having ability to copy an array with copy assignment operator and with copy ctor , both.
Also has a fantastic ability to move the resources ( array on heap) instead of copying it with move assignment operator and move ctor, when needed.
Unlike traditional arrays, the DynamicArray ADT can resize itself automatically. This means you can add or remove elements without worrying about the initial size or running out of space.
The array is allocated on the heap, allowing for efficient memory management. This ensures that the array can grow or shrink as needed, without the constraints of stack memory.
Once the DynamicArray goes out of scope, it is automatically removed from the heap. This eliminates the need for manual memory management, reducing the risk of memory leaks and simplifying your code.
Having an attractive behaviour with >> and << operator ie. you can use these operators with it like with primitive types.